requirements.txt
torch 
torchvision 
torchaudio 
matplotlib 
pillow 
numpy 
tqdm
fastapi
uvicorn
requests
python-multipart